Update service-dataflows-directquery.md
The changes proposed above are because we have had several customers creating tickets because the document does not goes over step by step on how to connect using direct query mode to a dataflow. The document just explains how to enable it but it does not go over the steps that need to be follow in desktop. Which should be imply but customers are not understanding so it is better to correct that right away and avoid more tickets